[refactor] centralize redis logic, create new module for job logging
