+++ snuba/manual_jobs/job_logging.py
@@ -1,0 +1,10 @@
+from logging import Logger
+
+
+class _MultiplexingRedisLogger:
+    def __init__(self, *loggers: Logger):
+        self.loggers = loggers
+
+    def log(self, *args, **kwargs):
+        for logger in self.loggers:
+            logger.log(*args, **kwargs)
